+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Rules/Utxow.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
@@ -60,12 +61,11 @@ import Cardano.Ledger.BaseTypes (
   ShelleyBase,
   StrictMaybe (..),
   quorum,
-  strictMaybeToMaybe,
  )
 import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.Core
-import Cardano.Ledger.Credential (Credential (KeyHashObj))
+import Cardano.Ledger.Credential (Credential (KeyHashObj), credKeyHashWitness)
 import Cardano.Ledger.Crypto (DSIGN, HASH)
 import Cardano.Ledger.Keys (GenDelegs, KeyHash, KeyRole (..), asWitness)
 import Cardano.Ledger.PoolParams (ppOwners)
@@ -82,8 +82,7 @@ import Cardano.Ledger.Shelley.Rules (
   validateNeededWitnesses,
  )
 import qualified Cardano.Ledger.Shelley.Rules as Shelley
-import Cardano.Ledger.Shelley.Tx (TxIn (..), extractKeyHashWitnessSet)
-import Cardano.Ledger.Shelley.TxCert (delegCWitness, requiresVKeyWitness, pattern ShelleyTxCertDeleg)
+import Cardano.Ledger.Shelley.Tx (TxIn (..))
 import Cardano.Ledger.Shelley.UTxO (ShelleyScriptsNeeded (..))
 import Cardano.Ledger.UTxO (EraUTxO (..), UTxO (..), txinLookup)
 import Control.Monad.Trans.Reader (asks)
@@ -448,33 +447,29 @@ witsVKeyNeeded utxo' tx genDelegs =
     wdrlAuthors :: Set (KeyHash 'Witness (EraCrypto era))
     wdrlAuthors = Map.foldrWithKey' accum Set.empty (unWithdrawals (txBody ^. withdrawalsTxBodyL))
       where
-        accum key _ = Set.union (extractKeyHashWitnessSet [getRwdCred key])
+        accum key _ !ans =
+          case credKeyHashWitness (getRwdCred key) of
+            Nothing -> ans
+            Just vkeyWit -> Set.insert vkeyWit ans
     owners :: Set (KeyHash 'Witness (EraCrypto era))
     owners = foldr' accum Set.empty (txBody ^. certsTxBodyL)
       where
-        accum (TxCertPool (RegPool pool)) ans =
+        accum (TxCertPool (RegPool pool)) !ans =
           Set.union
             (Set.map asWitness (ppOwners pool))
             ans
         accum _cert ans = ans
-    cwitness (ShelleyTxCertDeleg dc) = extractKeyHashWitnessSet [delegCWitness dc]
-    cwitness (TxCertPool pc) = extractKeyHashWitnessSet [poolCWitness pc]
-    cwitness (TxCertGenesis c) = Set.singleton (asWitness $ genesisCWitness c)
-    cwitness c = error $ show c ++ " does not have a witness"
-    -- key reg requires no witness but this is already filtered out by requiresVKeyWitness
-    -- before the call to `cwitness`, so this error should never be reached.
-
     certAuthors :: Set (KeyHash 'Witness (EraCrypto era))
     certAuthors = foldr' accum Set.empty (txBody ^. certsTxBodyL)
       where
-        accum cert ans | requiresVKeyWitness cert = Set.union (cwitness cert) ans
-        accum _cert ans = ans
+        accum cert !ans =
+          case getVKeyWitnessTxCert cert of
+            Nothing -> ans
+            Just vkeyWit -> Set.insert vkeyWit ans
     updateKeys :: Set (KeyHash 'Witness (EraCrypto era))
     updateKeys =
       asWitness
-        `Set.map` propWits
-          (strictMaybeToMaybe $ txBody ^. updateTxBodyG)
-          genDelegs
+        `Set.map` propWits (txBody ^. updateTxBodyG) genDelegs
 
 extSymmetricDifference :: (Ord k) => [a] -> (a -> k) -> [b] -> (b -> k) -> ([a], [b])
 extSymmetricDifference as fa bs fb = (extraA, extraB)